requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-ckeditor==6.4.2
django-dotenv==1.4.2
django-js-asset==2.0.0
Pillow==9.2.0
psycopg2==2.9.3
sqlparse==0.4.2
